techpack: display: Trigger panel reset after first kickoff on non-60 Hz
If we're using a refresh rate different from what the bootloader
initializes the display with, a panel reset is necessary for our custom
refresh rate to take effect.

After the first kickoff, request a panel reset from userspace using the
existing ESD recovery mechanism to update the refresh rate. Note that
this is only done if the target refresh rate is *not* 60 Hz because this
is redundant if we don't need a refresh rate update.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: I2a4dcae2043d651926929aa422ab34d655c15991
Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>
Signed-off-by: Diaz1401 <reagor8161@outlook.com>